CORTX:33336 MOved version test to test_data_pod_restart_api
